Slab allocator: make frees O(1)
This comes at the cost of an extra byte per object. This is not very
space efficient when compared to the SLUB allocator, which has no
overhead per object (the freelist is stored in-place, and even the
slab descriptor is stored in an existing data structure, the `struct
page` array). I wrote up some comments about the SLUB allocator in the
header comment in slab.h.

For the purposes of an educational OS, I think that the current slab
allocator design is "good enough". This does also have the benefit
over the SLUB allocator of allowing for the caching of initialized
objects. And if we want to implement the SLUB allocator, much of the
logic can be shared with the existing implementation.