Slab allocator (#8)
* Slab allocator: initial skeleton

* Implement logic to determine number of elements per slab

* Continue implementing slab allocator

* Add some unit tests for slab allocator

* Refactor physical memory manager to take a phys_rra argument

This clears the assumption that the main physical memory page
allocator will be used, paving the way for unit testing.

Also do some minor cleaning up, e.g., adding `const` to declarations
in the page/slab allocators where possible.

* Expose `slab_*()` and `slab_cache_*()` methods

For future unit testing

* Specify physical page allocator for slab allocator

This change is in anticipation of unit testing the slab allocator

* Slab allocator: make frees O(1)

This comes at the cost of an extra byte per object. This is not very
space efficient when compared to the SLUB allocator, which has no
overhead per object (the freelist is stored in-place, and even the
slab descriptor is stored in an existing data structure, the `struct
page` array). I wrote up some comments about the SLUB allocator in the
header comment in slab.h.

For the purposes of an educational OS, I think that the current slab
allocator design is "good enough". This does also have the benefit
over the SLUB allocator of allowing for the caching of initialized
objects. And if we want to implement the SLUB allocator, much of the
logic can be shared with the existing implementation.

* Add unit tests for physical RR allocator

* Miscellaneous test cleanup

* Various improvements to slab allocator

- Implement `struct list_head` for slab cache freelists. This probably will save a lot of headache in the future (I had a linked list debugging issue that bothered me for too long).
- Implement `ilog2` and `ilog2ceil` for computing orders (log base 2) from a decimal number.
- Create a test fixture for slab allocation. This includes writing `slab_cache_destroy()` and `slab_destroy()`, which are important but introduced me to many bugs.
- Write some unit tests for the slab allocator.

This commit is about half finished, I just want to save my work here. Still have to write some unit tests and do a lot of cleanup.

* More bugfixes and implementing tests for slab allocator

* Slab allocator and linked list tests

* Cleanup and update README with testing usage