Create LICENSE
Limine has a separate license, which is located in the root of the submodule. I don't think I need to include it here (https://opensource.stackexchange.com/questions/11563/should-i-include-the-license-of-packages-used-for-a-personal-project).

I want to license this permissively, as this is a personal project meant for my own (and perhaps others') educational purposes.